+++ database/nostr-sqlite/CHANGELOG.md
@@ -27,6 +27,12 @@
 
 -->
 
+## Unreleased
+
+### Performance
+
+- Optimize negentropy item queries
+
 ## v0.45.0 - 2026/08/05
 
 First release.